refactor: rename discord -> discord_bot
