+++ discord_bot/.gitignore
