+++ discord_bot/discord_bot/__init__.py
@@ -1,0 +1,1 @@
+"""Discord discord_bot."""
