+++ discord_bot/discord_bot/llm.py
