+++ discord_bot/prompts/sheep-duck-llama2-70B.txt
